Fix a lot of warnings due to reversed comparisons
This silences warnings when compiling with clang-10 and -std=c++2a, that
complain about various ambiguities in comparison operators operator==
and operator!= of iterator classes that inherit with CRTP from
iterator_adaptor_base.

This new set of ambiguities is introduced by C++20 rules for rewriting
comparison operators, that are designed to reduce the number of
comparison that you typically have to write, allowing the compiler to
rewrite and reverse comparison operators, so that most comparison
operators can be rewritten in terms of the base ones.

Turns out this creates some problems with CRTP because comparison
operators of the CRPT base, that were fine until C++17, now become
ambiguous.

In particular, this is due to the fact that for C++20 operator== and
operator!= are supposed to be symmetric (LHS and RHS need to be
swappable), and having

  CRTPBase::operator==(const CRTPDerived &D) const;

screws things up, because the type of the operator parameter
(CRTPDerived) is not equal to the type of the class for which we are
implementing the operator (CRTPBase).

This forces us to add the following

bool CRTPDerived::operator==(const CRTPDerived &D) const {
  return CRTBase::operator==(D)
}

for most of the classes that inherit from iterator_adaptor_base with
CRTP, to stop the compiler from complaining.